Fixes: [scm/git] - bail out to master branch as quickly as possible. This ensures that   even if something goes wrong along the deploy, you can fix the   errors and proceed one more time. Otherwise git fails to checkout   branch because you're in a 'detached state'. - change cleanup to finish [templates/BasicDeployment] - move the scp and extracting part to the template rather than scm.

+++ scm/git.py
@@ -3,7 +3,6 @@ from fabric.operations import local, run
 from fabric.context_managers import lcd, settings
 from os.path import exists
 from hashlib import md5
-from commands import scp
 
 
 class Backend:
@@ -11,7 +10,7 @@ class Backend:
         self.deploy = deploy
         self.url = url
         self.config = config
-        self.deploy.add_signal_handler('finish', self.cleanup)
+        self.deploy.add_signal_handler('finish', self.finish)
         self.branch = self.deploy.branch or 'master'
 
     def refresh_repo(self):
@@ -30,37 +29,57 @@ class Backend:
         #file is not on the server. This basically means, that there was no
         #deploy yet and we need to make a fresh one.
         self.removed_files_list = None
-        with lcd("repo"):
-            head = local("git checkout %(branch)s && git rev-parse HEAD" % {
-                'branch': self.branch,
-            }, capture=True)
-            self.payload_file = 'payload-%s.tar.gz' % head
+        self.payload_file = None
+
         with settings(warn_only=True):
             sha1 = run("cat %(project)s/.last-deployment" % {
                 'project': self.deploy.project_path
             })
-            if sha1.failed:
-                #there was no last deployment info -> we need to deploy a full
-                #tree of the code.
-                with lcd("repo"):
-                    local("git checkout %(branch)s && git ls-files > ../payload" % {
-                        'branch': self.branch
-                    })
-                    local("tar -cf ../%(payload)s -T ../payload" % {
-                        'payload': self.payload_file
-                    })
-                    
-            self.deploy.emit_signal("pre-copy")
-            scp(self.payload_file, "/tmp")
-            self.deploy.emit_signal("pre-extract")
-            self.deploy.emit_signal("post-extract")
+            with lcd("repo"):
+                local("git checkout %s" % self.branch)
+                self.head = local("git checkout %(branch)s && git rev-parse HEAD" % {
+                    'branch': self.branch,
+                }, capture=True)
+                self.payload_file = 'payload-%s.tar.gz' % self.head
+                if sha1.failed:
+                    #there was no last deployment info -> we need to deploy a full
+                    #tree of the code.
+                        local("git ls-files > ../payload" % {
+                            'branch': self.branch
+                        })
+                        local("tar -czf ../%(payload)s -T ../payload" % {
+                            'payload': self.payload_file
+                        })
+                else:
+                    if self.head != sha1:
+                    #we prepare the diff
+                        local("git diff --name-status --no-renames %s |\
+                                   egrep '^(A|M)' | cut -f 2 > ../payload" % sha1)
+                        local("git diff --name-status --no-renames %s |\
+                                   egrep '^D' | cut -f 2 > ../remove-files" % sha1)
+                    else:
+                        self.payload_file = None
+                #just in case enything goes berserk in the future
+                #we bail out to master branch, so future deploys
+                #could resume sanely.
+                local("git checkout master")
+                
+            if self.payload_file:
+                local("tar -czf ../%(payload)s -T ../payload" % {
+                    'payload': self.payload_file
+                })
 
+                            
         return (self.payload_file, self.removed_files_list)
 
-    def cleanup(self):
-        with lcd("repo"):
-            local("git checkout master")
-        local("rm -f payload %(payload)s %(remove)s" % {
-            'payload': self.payload_file,
-            'remove': self.removed_files_list
+    def finish(self):
+        if self.payload_file:
+            local("rm -f payload %(payload)s %(remove)s" % {
+                'payload': self.payload_file,
+                'remove': self.removed_files_list
+            })
+            run("rm -f /tmp/" + self.payload_file)
+        run("echo %(head)s > %(project)s/.last-deployment" % {
+            "head": self.head,
+            "project": self.deploy.project_path
         })
